Extend rollback coverage, tidy output helpers, and record testing ADR
Code:
- output/mod.rs: remove the empty `if delta.death` placeholder branch from
  `apply_one_health_delta`.
- state: extract the shared reset in `begin_frame_rollback` and
  `commit_frame_tracking` into a private `clear_frame_rollback` helper. Split
  the module into `state/mod.rs` + `state/tests.rs` to stay within the
  module-line budget; the three basic tests now use the shared `#[from(state)]`
  fixture.
- output/tests: `prime_state` uses the `push_velocity_input` helper; the
  negative-weight velocity test is now two-phase (asserts a +1 weight IS
  applied for an unsupported entity before asserting a -1 weight is skipped),
  so it fails if velocity output is never emitted.
- output/tests/failure_paths: add `failed_step_rolls_back_applied_unsequenced`,
  a full-pipeline sibling that drives an unsequenced `DamageEvent` through the
  `DamageInbox`, forces a step error, and asserts `applied_unsequenced` is
  restored to its pre-frame value via the real cache-to-step path.

Docs:
- dbsp-synchronization-guide.md: document that `expected_health_retractions`
  is transient (cleared/rebuilt each cache pass, read only within the frame)
  and therefore excluded from rollback, reconciling the mutated-state list with
  the rollback methods; reword §6 to stop claiming the bounded `rstest`
  matrices are "equivalent" to property tests.
- Add docs/adr-003-bounded-rstest-over-property-testing.md recording the
  bounded-exhaustive-`rstest`-over-property-testing decision as the documented,
  approved exception; index it in docs/README.md.
